Get a layer from a web coverage service within a bounding box

The function sends a query to a WCS service, downloads it to a temporary file from which it is read with terra::rast() - if needed reprojected - and returned as a SpatRasterobject

get_coverage_wcs(
  wcs = c("dtm", "dsm", "omz", "omw"),
  bbox,
  layername,
  resolution,
  wcs_crs = "EPSG:4258",
  output_crs = "EPSG:31370",
  bbox_crs = "EPSG:31370",
  version = c("1.0.0", "2.0.1"),
  ...
)

Arguments

wcs

One of "dtm", "dsm", "omz", "omw"

bbox

An object of class bbox of length 4.

layername

Character string; name of the layer

resolution

Output resolution in meters

wcs_crs

Native CRS in which the raster layers are stored on the WCS

output_crs

Output CRS. May involve reprojection.

bbox_crs

CRS in which bbox coordinates are passed

version

WCS version to be used.

...

Additional key-value pairs passed on to the WCS query

Value

A SpatRaster object

Details

The following WCS services can currently be used:

"omz": orthophotomosaic summer images Flanders

"omw": orthophotomosaic winter images Flanders

"dtm": digital terrain model Flanders

"dsm": digital surface model Flanders See metadata Vlaanderen for more information # nolint: line_length_linter.

See also

Examples

if (FALSE) {
bbox <- sf::st_bbox(
  c(xmin = 155800, xmax = 155850, ymin = 132700, ymax = 132750),
  crs = sf::st_crs(31370)
)
get_coverage_wcs(
  wcs = "dsm",
  bbox = bbox,
  layername = "EL.GridCoverage.DSM",
  resolution = 1
)
}