inbospatial: A Collection of Useful R Functions for Spatial Data

Useful R functions developed at the Research Institute for Nature and Forest (INBO) for dealing with spatial raster or vector data.

See also

Author

Maintainer: Hans Van Calster hans.vancalster@inbo.be (ORCID) (Research Institute for Nature and Forest (INBO))

Other contributors:

Floris Vanderhaeghe floris.vanderhaeghe@inbo.be (ORCID) (Research Institute for Nature and Forest (INBO)) [contributor]

Research Institute for Nature and Forest (INBO) info@inbo.be [copyright holder, funder]