addTiles() wrapper functions for WMTS services

The add_wmts_*() functions are shorthand alternatives for a fully specified leaflet::addTiles() statement. Appropriate attribution is added to the Leaflet map depending on the layer.

Usage

add_wmts_nl_brt(map, layer = c("standaard", "grijs", "pastel", "water"), ...)

add_wmts_nl_ortho(map, ...)

Arguments

map

a map widget object created from leaflet()

layer

String that defines which layer to use, if the function supports more than one. In such case, the available strings are shown in the Usage section. The first value is used if missing.

...

Further arguments passed to leaflet::addTiles().

Value

A leaflet HTML widget object.

Details

On condition that these functions continue being maintained, using the shorthand functions should make your scripts more futureproof.

Abbreviations

Prefixes

be_

Belgium

nl_

Netherlands

Suffixes

_ortho

Orthophotographs

_brt

Topographic BRT layers (Netherlands; 'Basisregistratie Topografie')

See also

Author

Floris Vanderhaeghe, https://github.com/florisvdh

Examples

library(leaflet)
leaflet() |>
  setView(lng = 5.5, lat = 52.5, zoom = 9) |>
  add_wmts_nl_brt()

leaflet() |>
  setView(lng = 5.4, lat = 52.2, zoom = 14) |>
  add_wmts_nl_ortho()